+++ core/clients/key_flow_continuous_refresh.go
@@ -116,7 +116,7 @@ func (refresher *continuousTokenRefresher) waitUntilTimestamp(timestamp time.Tim
 //   - (false, nil) if not successful but should be retried.
 //   - (_, err) if not successful and shouldn't be retried.
 func (refresher *continuousTokenRefresher) refreshToken() (bool, error) {
-	err := refresher.flow.RefreshAccessToken()
+	err := refresher.flow.refreshAccessToken()
 	if err == nil {
 		return true, nil
 	}